src/main/java/com/outgrowthsolutions/ogsrecipeapp/controllers/IndexController.java
package com.outgrowthsolutions.ogsrecipeapp.controllers;

import com.outgrowthsolutions.ogsrecipeapp.services.RecipeService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;

@Slf4j
@RequestMapping("/")
@Controller
public class IndexController {
    private final RecipeService recipeService;

    public IndexController(RecipeService recipeService) {
        this.recipeService = recipeService;
    }


    @RequestMapping({"","index"})
    public String viewIndex(Model model){
        log.debug("IndexController::viewIndex");
       model.addAttribute("recipes",recipeService.getRecipes());
        return "index";
    }
}
